+++ routes/departments.js
@@ -1,76 +1,14 @@
 const express = require('express');
 const router = express.Router();
-const { Dept } = require('../models/Departement');
 
-const bodyParser = require('body-parser');
-router.use(bodyParser.json());
-router.use(express.json());
+const { add, edit, del, getAll } = require('../controllers/depts_controlles')
 
-const logger = require('../logger/logger');
-//const loggerErr = require('../logger/logger-err')
+router.post('/add', add)
 
-const expressWinston = require('express-winston');
+router.patch('/edit/:id', edit)
 
-router.use(expressWinston.logger({
-   winstonInstance: logger,
-   statusLevels: true
-}))
+router.delete('/delete/:id', del)
 
-router.post('/add', async (req,res) => {
-
-    const edit = req.body.edit 
-    try
-    {
-      const dept = new Dept({
-        name: edit
-      })
-    
-      await dept.save();
-      return res.status(201).json({ error: false, message: "Succes: Department have been added" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
-
-router.patch('/edit/:id', async (req,res) => {
-
-    try
-    {
-    const id = req.params.id 
-    const edit = req.body.name
-    const dept = await Dept.findById(id, {name: edit })
-    
-    return res.status(201).json({ error: false, message: "Succes: Department have been edited" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
-
-router.delete('/delete/:id', async (req,res) => {
-
-    try
-    {
-    const id = req.params.id 
-    const dept = await Dept.findById(id).remove()
-    
-    return res.status(201).json({ error: false, message: "Succes: Department have been deleted" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
-
-router.get('/getAll', async (req,res) => {
-
-    try
-    {
-    const dept = await Dept.find()
-    
-    return res.status(201).json({ error: false, message: "Succes: Department have been geted All" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
-
-//router.use(expressWinston.errorLogger(loggerErr))
+router.get('/getAll', getAll)
 
 module.exports = router;
