+++ routes/localisations.js
@@ -1,47 +1,11 @@
 const express = require('express');
 const router = express.Router();
-const  Localisation  = require('../models/Localisation');
+const { add, del, getAll } = require('../controllers/loc_controllers')
 
-router.post('/add', async (req,res) => {
+router.post('/add', add)
 
-    const edit = req.body.edit 
-    try
-    {
-      const loc = new Localisation({
-        name: edit
-      })
-    
-      await loc.save();
-      return res.status(201).json({ error: false, message: "Succes: Location have been added" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
+router.delete('/delete/:id', del)
 
-router.delete('/delete/:id', async (req,res) => {
-
-    try
-    {
-      const id = req.params.id 
-      const loc = await Localisation.findById(id).remove()
-    
-      return res.status(201).json({ error: false, message: "Succes: Location have been deleted" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
-
-router.get('/getAll', async (req,res) => {
-
-    try
-    {
-    const loc = await Localisation.find()
-    console.log(loc)
-    
-    return res.status(201).json({ error: false, message: "Succes: Locations have been Geted" });
-   } catch (err) {
-      return res.status(500).json({ error: true, message: "Internal Servor Error" });
-   }
-})
+router.get('/getAll', getAll)
 
 module.exports = router;
