+++ utils/validationSchema.js
@@ -1,16 +1,36 @@
 //Joi
 //Joi-password complexity
 const Joi = require('joi');
-const passwordComplexity = require('joi-password-complexity');
 
-const logInBodyValidation = (body) => {
-    const schema = Joi.object({
-        username: Joi.string().email().required().label("User Name"),
-        password: passwordComplexity().required().label("Password")
-    })
-    return schema.validate(body);
-}
+const logInBodyValidation = Joi.object({
+    username: Joi.string().email().lowercase().required(),
+    password: Joi.string().min(4).max(15).required()
+})
+const registerEmpValidation = Joi.object({
+    email: Joi.string().email().lowercase().required(),
+    password: Joi.string().min(4).max(15).required(),
+    firstN: Joi.string().required(),
+    lastN: Joi.string().required(),
+    role: Joi.string().required(),
+    sin: Joi.string().required(),
+    dob: Joi.string().required(),
+    phoneN: Joi.string().required(),
+    image: Joi.string().required(),
+    localisation: Joi.string().required(),
+    dept: Joi.string().required()
+})
 
-module.exports = {
-    logInBodyValidation
-}
+const registerUsersValidation = Joi.object({
+    email: Joi.string().email().lowercase().required(),
+    password: Joi.string().min(4).max(15).required(),
+    firstN: Joi.string().required(),
+    lastN: Joi.string().required(),
+    dob: Joi.string().required(),
+    image: Joi.string().required(),
+    phoneN: Joi.string().required(),
+    localisation: Joi.string().required(),
+    dept: Joi.string().required(),
+    tokens: Joi.string()
+})
+
+module.exports = { logInBodyValidation, registerUsersValidation }
